+++ gdal/alg/polygonize.cpp
@@ -126,13 +126,13 @@ void RPolygon::Coalesce()
                 if( anBase[anBase.size()-2] == anString[0]
                     && anBase[anBase.size()-1] == anString[1] )
                 {
-                    Merge( iBaseString, static_cast<int>(iString), 1 );
+                    Merge( static_cast<int>(iBaseString), static_cast<int>(iString), 1 );
                     bMergeHappened = TRUE;
                 }
                 else if( anBase[anBase.size()-2] == anString[anString.size()-2]
                          && anBase[anBase.size()-1] == anString[anString.size()-1] )
                 {
-                    Merge( iBaseString, static_cast<int>(iString), -1 );
+                    Merge( static_cast<int>(iBaseString), static_cast<int>(iString), -1 );
                     bMergeHappened = TRUE;
                 }
             }
